+++ src/main/java/com/nvidia/spark/rapids/jni/kudo/SlicedBufferSerializer.java
@@ -43,168 +43,168 @@ import static com.nvidia.spark.rapids.jni.kudo.KudoSerializer.padForHostAlignmen
  * </p>
  */
 class SlicedBufferSerializer implements HostColumnsVisitor<Void> {
-    private final SliceInfo root;
-    private final BufferType bufferType;
-    private final DataWriter writer;
+  private final SliceInfo root;
+  private final BufferType bufferType;
+  private final DataWriter writer;
 
-    private final Deque<SliceInfo> sliceInfos = new ArrayDeque<>();
-    private long totalDataLen;
+  private final Deque<SliceInfo> sliceInfos = new ArrayDeque<>();
+  private long totalDataLen;
 
-    SlicedBufferSerializer(long rowOffset, long numRows, BufferType bufferType, DataWriter writer) {
-        this.root = new SliceInfo(rowOffset, numRows);
-        this.bufferType = bufferType;
-        this.writer = writer;
-        this.sliceInfos.addLast(root);
-        this.totalDataLen = 0;
+  SlicedBufferSerializer(long rowOffset, long numRows, BufferType bufferType, DataWriter writer) {
+    this.root = new SliceInfo(rowOffset, numRows);
+    this.bufferType = bufferType;
+    this.writer = writer;
+    this.sliceInfos.addLast(root);
+    this.totalDataLen = 0;
+  }
+
+  public long getTotalDataLen() {
+    return totalDataLen;
+  }
+
+  @Override
+  public Void visitStruct(HostColumnVectorCore col, List<Void> children) {
+    SliceInfo parent = sliceInfos.peekLast();
+
+    try {
+      switch (bufferType) {
+        case VALIDITY:
+          totalDataLen += this.copySlicedValidity(col, parent);
+          return null;
+        case OFFSET:
+        case DATA:
+          return null;
+        default:
+          throw new IllegalArgumentException("Unexpected buffer type: " + bufferType);
+      }
+
+    } catch (IOException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
+  @Override
+  public Void preVisitList(HostColumnVectorCore col) {
+    SliceInfo parent = sliceInfos.getLast();
+
+
+    long bytesCopied = 0;
+    try {
+      switch (bufferType) {
+        case VALIDITY:
+          bytesCopied = this.copySlicedValidity(col, parent);
+          break;
+        case OFFSET:
+          bytesCopied = this.copySlicedOffset(col, parent);
+          break;
+        case DATA:
+          break;
+        default:
+          throw new IllegalArgumentException("Unexpected buffer type: " + bufferType);
+      }
+
+    } catch (IOException e) {
+      throw new RuntimeException(e);
     }
 
-    public long getTotalDataLen() {
-        return totalDataLen;
+    SliceInfo current;
+    if (col.getOffsets() != null) {
+      long start = col.getOffsets()
+          .getInt(parent.offset * Integer.BYTES);
+      long end = col.getOffsets().getInt((parent.offset + parent.rowCount) * Integer.BYTES);
+      long rowCount = end - start;
+
+      current = new SliceInfo(start, rowCount);
+    } else {
+      current = new SliceInfo(0, 0);
     }
 
-    @Override
-    public Void visitStruct(HostColumnVectorCore col, List<Void> children) {
-        SliceInfo parent = sliceInfos.peekLast();
+    sliceInfos.addLast(current);
 
-        try {
-            switch (bufferType) {
-                case VALIDITY:
-                    totalDataLen += this.copySlicedValidity(col, parent);
-                    return null;
-                case OFFSET:
-                case DATA:
-                    return null;
-                default:
-                    throw new IllegalArgumentException("Unexpected buffer type: " + bufferType);
-            }
+    totalDataLen += bytesCopied;
+    return null;
+  }
 
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
+  @Override
+  public Void visitList(HostColumnVectorCore col, Void preVisitResult, Void childResult) {
+    sliceInfos.removeLast();
+    return null;
+  }
+
+  @Override
+  public Void visit(HostColumnVectorCore col) {
+    SliceInfo parent = sliceInfos.getLast();
+    try {
+      switch (bufferType) {
+        case VALIDITY:
+          totalDataLen += this.copySlicedValidity(col, parent);
+          return null;
+        case OFFSET:
+          totalDataLen += this.copySlicedOffset(col, parent);
+          return null;
+        case DATA:
+          totalDataLen += this.copySlicedData(col, parent);
+          return null;
+        default:
+          throw new IllegalArgumentException("Unexpected buffer type: " + bufferType);
+      }
+    } catch (IOException e) {
+      throw new RuntimeException(e);
     }
+  }
 
-    @Override
-    public Void preVisitList(HostColumnVectorCore col) {
-        SliceInfo parent = sliceInfos.getLast();
+  private long copySlicedValidity(HostColumnVectorCore column, SliceInfo sliceInfo) throws IOException {
+    if (column.getValidity() != null && sliceInfo.getRowCount() > 0) {
+      HostMemoryBuffer buff = column.getValidity();
+      long len = sliceInfo.getValidityBufferInfo().getBufferLength();
+      writer.copyDataFrom(buff, sliceInfo.getValidityBufferInfo().getBufferOffset(),
+          len);
+      return padForHostAlignment(writer, len);
+    } else {
+      return 0;
+    }
+  }
 
+  private long copySlicedOffset(HostColumnVectorCore column, SliceInfo sliceInfo) throws IOException {
+    if (sliceInfo.rowCount <= 0 || column.getOffsets() == null) {
+      // Don't copy anything, there are no rows
+      return 0;
+    }
+    long bytesToCopy = (sliceInfo.rowCount + 1) * Integer.BYTES;
+    long srcOffset = sliceInfo.offset * Integer.BYTES;
+    HostMemoryBuffer buff = column.getOffsets();
+    writer.copyDataFrom(buff, srcOffset, bytesToCopy);
+    return padForHostAlignment(writer, bytesToCopy);
+  }
 
-        long bytesCopied = 0;
-        try {
-            switch (bufferType) {
-                case VALIDITY:
-                    bytesCopied = this.copySlicedValidity(col, parent);
-                    break;
-                case OFFSET:
-                    bytesCopied = this.copySlicedOffset(col, parent);
-                    break;
-                case DATA:
-                    break;
-                default:
-                    throw new IllegalArgumentException("Unexpected buffer type: " + bufferType);
-            }
+  private long copySlicedData(HostColumnVectorCore column, SliceInfo sliceInfo) throws IOException {
+    if (sliceInfo.rowCount > 0) {
+      DType type = column.getType();
+      if (type.equals(DType.STRING)) {
+        long startByteOffset = column.getOffsets().getInt(sliceInfo.offset * Integer.BYTES);
+        long endByteOffset = column.getOffsets().getInt((sliceInfo.offset + sliceInfo.rowCount) * Integer.BYTES);
+        long bytesToCopy = endByteOffset - startByteOffset;
+        if (column.getData() == null) {
+          if (bytesToCopy != 0) {
+            throw new IllegalStateException("String column has no data buffer, " +
+                "but bytes to copy is not zero: " + bytesToCopy);
+          }
 
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        SliceInfo current;
-        if (col.getOffsets() != null) {
-            long start = col.getOffsets()
-                    .getInt(parent.offset * Integer.BYTES);
-            long end = col.getOffsets().getInt((parent.offset + parent.rowCount) * Integer.BYTES);
-            long rowCount = end - start;
-
-            current = new SliceInfo(start, rowCount);
+          return 0;
         } else {
-            current = new SliceInfo(0, 0);
+          writer.copyDataFrom(column.getData(), startByteOffset, bytesToCopy);
+          return padForHostAlignment(writer, bytesToCopy);
         }
-
-        sliceInfos.addLast(current);
-
-        totalDataLen += bytesCopied;
-        return null;
-    }
-
-    @Override
-    public Void visitList(HostColumnVectorCore col, Void preVisitResult, Void childResult) {
-        sliceInfos.removeLast();
-        return null;
-    }
-
-    @Override
-    public Void visit(HostColumnVectorCore col) {
-        SliceInfo parent = sliceInfos.getLast();
-        try {
-            switch (bufferType) {
-                case VALIDITY:
-                    totalDataLen += this.copySlicedValidity(col, parent);
-                    return null;
-                case OFFSET:
-                    totalDataLen += this.copySlicedOffset(col, parent);
-                    return null;
-                case DATA:
-                    totalDataLen += this.copySlicedData(col, parent);
-                    return null;
-                default:
-                    throw new IllegalArgumentException("Unexpected buffer type: " + bufferType);
-            }
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private long copySlicedValidity(HostColumnVectorCore column, SliceInfo sliceInfo) throws IOException {
-        if (column.getValidity() != null && sliceInfo.getRowCount() > 0) {
-            HostMemoryBuffer buff = column.getValidity();
-            long len = sliceInfo.getValidityBufferInfo().getBufferLength();
-            writer.copyDataFrom(buff, sliceInfo.getValidityBufferInfo().getBufferOffset(),
-                    len);
-            return padForHostAlignment(writer, len);
-        } else {
-            return 0;
-        }
-    }
-
-    private long copySlicedOffset(HostColumnVectorCore column, SliceInfo sliceInfo) throws IOException {
-        if (sliceInfo.rowCount <= 0 || column.getOffsets() == null) {
-            // Don't copy anything, there are no rows
-            return 0;
-        }
-        long bytesToCopy = (sliceInfo.rowCount + 1) * Integer.BYTES;
-        long srcOffset = sliceInfo.offset * Integer.BYTES;
-        HostMemoryBuffer buff = column.getOffsets();
-        writer.copyDataFrom(buff, srcOffset, bytesToCopy);
+      } else if (type.getSizeInBytes() > 0) {
+        long bytesToCopy = sliceInfo.rowCount * type.getSizeInBytes();
+        long srcOffset = sliceInfo.offset * type.getSizeInBytes();
+        writer.copyDataFrom(column.getData(), srcOffset, bytesToCopy);
         return padForHostAlignment(writer, bytesToCopy);
+      } else {
+        return 0;
+      }
+    } else {
+      return 0;
     }
-
-    private long copySlicedData(HostColumnVectorCore column, SliceInfo sliceInfo) throws IOException {
-        if (sliceInfo.rowCount > 0) {
-            DType type = column.getType();
-            if (type.equals(DType.STRING)) {
-                long startByteOffset = column.getOffsets().getInt(sliceInfo.offset * Integer.BYTES);
-                long endByteOffset = column.getOffsets().getInt((sliceInfo.offset + sliceInfo.rowCount) * Integer.BYTES);
-                long bytesToCopy = endByteOffset - startByteOffset;
-                if (column.getData() == null) {
-                    if (bytesToCopy != 0) {
-                        throw new IllegalStateException("String column has no data buffer, " +
-                                "but bytes to copy is not zero: " + bytesToCopy);
-                    }
-
-                    return 0;
-                } else {
-                    writer.copyDataFrom(column.getData(), startByteOffset, bytesToCopy);
-                    return padForHostAlignment(writer, bytesToCopy);
-                }
-            } else if (type.getSizeInBytes() > 0) {
-                long bytesToCopy = sliceInfo.rowCount * type.getSizeInBytes();
-                long srcOffset = sliceInfo.offset * type.getSizeInBytes();
-                writer.copyDataFrom(column.getData(), srcOffset, bytesToCopy);
-                return padForHostAlignment(writer, bytesToCopy);
-            } else {
-                return 0;
-            }
-        } else {
-            return 0;
-        }
-    }
+  }
 }
